# backend/openapi_temp.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZiggyAI Trading Platform",
    "description": "Complete AI-powered trading platform with market analysis, news, signals, and execution",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/crypto/quotes": {
      "get": {
        "summary": "Crypto Quotes",
        "operationId": "crypto_quotes_api_crypto_quotes_get",
        "parameters": [
          {
            "name": "symbols",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Comma-separated symbols like BTC-USD,ETH-USD,SOL-USD",
              "title": "Symbols"
            },
            "description": "Comma-separated symbols like BTC-USD,ETH-USD,SOL-USD"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/crypto/ohlc": {
      "get": {
        "summary": "Crypto Ohlc",
        "operationId": "crypto_ohlc_api_crypto_ohlc_get",
        "parameters": [
          {
            "name": "symbols",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Comma-separated symbols",
              "title": "Symbols"
            },
            "description": "Comma-separated symbols"
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(1m|5m)$",
              "default": "1m",
              "title": "Interval"
            }
          },
          {
            "name": "minutes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10080,
              "minimum": 1,
              "default": 240,
              "title": "Minutes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}